refactor: move reconcile.is_repair -> buf_state
